Move new glimmer style string to its own home
