+++ ui/app/components/chart-primitives/v-annotations.js
@@ -1,6 +1,7 @@
 import Component from '@glimmer/component';
 import { htmlSafe } from '@ember/template';
 import { action } from '@ember/object';
+import styleString from 'nomad-ui/utils/properties/glimmer-style-string';
 
 const iconFor = {
   error: 'cancel-circle-fill',
@@ -11,30 +12,6 @@ const iconClassFor = {
   error: 'is-danger',
   info: '',
 };
-
-// TODO: This is what styleStringProperty looks like in the pure decorator world
-function styleString(target, name, descriptor) {
-  if (!descriptor.get) throw new Error('styleString only works on getters');
-  const orig = descriptor.get;
-  descriptor.get = function() {
-    const styles = orig.apply(this);
-
-    let str = '';
-
-    if (styles) {
-      str = Object.keys(styles)
-        .reduce(function(arr, key) {
-          const val = styles[key];
-          arr.push(key + ':' + (typeof val === 'number' ? val.toFixed(2) + 'px' : val));
-          return arr;
-        }, [])
-        .join(';');
-    }
-
-    return htmlSafe(str);
-  };
-  return descriptor;
-}
 
 export default class ChartPrimitiveVAnnotations extends Component {
   @styleString
